+++ src/lib/components/EvmToCell.tsx
@@ -1,7 +1,7 @@
 import { Flex, Text } from "@chakra-ui/react";
 
-import type { Option } from "lib/types";
-import type { EvmToAddress } from "lib/utils";
+import type { EvmToAddress, HexAddr20, Option } from "lib/types";
+import { EvmMethodName } from "lib/types";
 
 import { ExplorerLink } from "./ExplorerLink";
 import { CustomIcon } from "./icon";
@@ -10,40 +10,52 @@ interface EvmToCellProps {
   toAddress: Option<EvmToAddress>;
 }
 
+const EvmToCellCreate = ({
+  address,
+}: {
+  address: HexAddr20;
+  evmTxHash?: string;
+}) => (
+  <Flex direction="column">
+    <Text variant="body3" color="text.disabled">
+      Created Contract
+    </Text>
+    {/* TODO: fix contract addresses */}
+    <Flex gap={1} align="center">
+      <CustomIcon name="contract-address" boxSize={3} color="primary.main" />
+      <ExplorerLink
+        value={address}
+        type="evm_contract_address"
+        showCopyOnHover
+      />
+    </Flex>
+  </Flex>
+);
+
 export const EvmToCell = ({ toAddress }: EvmToCellProps) => {
-  if (toAddress?.isCreatedContract)
+  if (!toAddress)
     return (
-      <Flex direction="column">
-        <Text variant="body3" color="text.disabled">
-          Created Contract
-        </Text>
-        <Flex gap={1} align="center">
-          <CustomIcon
-            name="contract-address"
-            boxSize={3}
-            color="primary.main"
-          />
-          <ExplorerLink
-            value={toAddress.address}
-            type={toAddress.type}
-            showCopyOnHover
-          />
-        </Flex>
-      </Flex>
+      <Text variant="body2" color="text.dark">
+        -
+      </Text>
     );
 
-  if (toAddress)
+  if (toAddress.toType === EvmMethodName.Create)
     return (
-      <ExplorerLink
-        value={toAddress.address}
-        type={toAddress.type}
-        showCopyOnHover
+      <EvmToCellCreate
+        address={toAddress.address}
+        evmTxHash={toAddress.evmTxHash}
       />
     );
 
+  if (toAddress.toType === EvmMethodName.CallErc20Factory)
+    return <EvmToCellCreate address={toAddress.address} />;
+
   return (
-    <Text variant="body2" color="text.dark">
-      -
-    </Text>
+    <ExplorerLink
+      value={toAddress.address}
+      type="user_address"
+      showCopyOnHover
+    />
   );
 };